+++ app/controllers/api/v1/base_controller.rb
@@ -1,2 +1,6 @@
+# Api::V1:BaseController serves as a parent controller (or ApplicationController if your prefer) for all other controllers found in the Api::V1 namespace.
+# 
+# @since 0.1.0
+# @author Robert Klubenspies
 class Api::V1::BaseController < ApplicationController
 end