+++ app/controllers/application_controller.rb
@@ -1,13 +1,7 @@
+# Methods found in the ApplicationController are available to all controllers that inherit from it.
+# 
+# @since 0.1.0
+# @author Robert Klubenspies
 class ApplicationController < ActionController::Base
   protect_from_forgery
-  # before_filter :check_cookies
-  # 
-  # def check_cookies
-  #   if cookies[:prayed_for] == nil
-  #     cookies[:prayed_for] = []
-  #   end
-  #   if cookies[:reported] == nil
-  #     cookies[:reported] = []
-  #   end
-  # end
 end