+++ app/controllers/legal_controller.rb
@@ -1,8 +1,17 @@
+# The LegalController handles/displays static pages of legal information (like the Terms of Service and Privacy Policy).
+# 
+# @since 0.1.0
+# @author Robert Klubenspies
 class LegalController < ApplicationController
+  # Displays Terms of Service screen
+  # @since 0.1.0
+  # @author Robert Klubenspies
   def tos
   end
-
+  
+  # Displays Privacy Policy screen
+  # @since 0.1.0
+  # @author Robert Klubenspies
   def privacy
   end
-
 end